[refactor] Move `GridItemCellRenderController` to single file
