+++ grid-cell/grid-cell/View/ViewController/GridViewController.swift
@@ -8,36 +8,6 @@
 import RxSwift
 import UIKit
 
-final class GridItemCellRenderController {
-  private var cell: GridItemCell? = nil
-  private let viewModel: GridItemCellViewModel
-  
-  init(viewModel: GridItemCellViewModel) {
-    self.viewModel = viewModel
-  }
-  
-  func setupCell(with cell: GridItemCell) {
-    self.cell = cell
-  }
-  
-  func preload() {
-    viewModel.loadImage { [weak self] result in
-      switch result {
-      case let .success(data):
-        self?.cell?.updateImage(with: data)
-      case .failure:
-        break
-      }
-    }
-    cell?.updateUI(with: viewModel)
-  }
-  
-  func cancelLoad() {
-    viewModel.cancelLoadImageTask()
-    cell = nil
-  }
-}
-
 final class GridViewModel {
   let gridLoader: GridLoader
   let imageLoader: GridImageDataLoader
